requirements.txt
django==2.1.3
gunicorn
django-heroku
djangorestframework==3.9.0